move menu_blink_system to new effects module
